Move user-list
