+++ client/src/app/users/user-list/user-list.component.html
@@ -5,49 +5,8 @@
       <mat-card-header>
         <mat-card-title class="user-list-title">Users</mat-card-title>
       </mat-card-header>
+      <app-user-filter [filteredUsers]="filteredUsers" (filteredUsersChange)="filteredUsers=$event"></app-user-filter>
       <mat-card-content>
-
-        <div class="flex-row gap-8 flex-wrap">
-          <!-- Examples of filtering in Angular -->
-
-          <mat-form-field class="input-field">
-            <mat-label>Name</mat-label>
-            <input matInput data-test="userNameInput" placeholder="Filter by name"
-            [(ngModel)]="userName" (input)="updateFilter()">
-            <mat-hint>Filtered on client</mat-hint>
-          </mat-form-field>
-
-          <mat-form-field class="input-field">
-            <mat-label>Company</mat-label>
-            <input matInput data-test="userCompanyInput" placeholder="Filter by company"
-            [(ngModel)]="userCompany" (input)="updateFilter()">
-            <mat-hint>Filtered on client</mat-hint>
-          </mat-form-field>
-        </div>
-
-        <div class="flex-row gap-8 flex-wrap">
-          <!-- Examples of filtering on the server -->
-
-          <mat-form-field class="input-field">
-            <mat-label>Age</mat-label>
-            <input matInput data-test="userAgeInput" type="number" placeholder="Filter by age"
-            min="0" max="200" [(ngModel)]="userAge" (input)="getUsersFromServer()">
-            <mat-hint>Filtered on server</mat-hint>
-          </mat-form-field>
-
-          <mat-form-field class="input-field">
-            <mat-label>Role</mat-label>
-            <mat-select (selectionChange)="getUsersFromServer()" [(ngModel)]="userRole" data-test="userRoleSelect">
-              <mat-option>--</mat-option>
-              <mat-option value="admin">Admin</mat-option>
-              <mat-option value="editor">Editor</mat-option>
-              <mat-option value="viewer">Viewer</mat-option>
-            </mat-select>
-            <mat-hint>Filtered on server</mat-hint>
-          </mat-form-field>
-        </div>
-
-        <br>
         <div class="flex-row gap-8 flex-wrap">
           <label>View type: </label>
           <mat-radio-group aria-label="View type" [(ngModel)]="viewType" data-test="viewTypeRadio">
@@ -55,20 +14,19 @@
             <mat-radio-button value="list">List</mat-radio-button>
           </mat-radio-group>
         </div>
-
       </mat-card-content>
     </mat-card>
   </div>
 </div>
 
 <div class="flex-row">
-  <div class="flex-1" *ngIf="serverFilteredUsers; else usersError" >
+  <div class="flex-1" *ngIf="this.filteredUsers; else usersError" >
     <!-- Switch between card and list view based on the viewType variable, set above in the mar-radio-group -->
     <div [ngSwitch]="viewType">
 
       <!-- Card grid view -->
       <div *ngSwitchCase="'card'" class="user-cards-container flex-row gap-8 flex-wrap">
-        <app-user-card [simple]="true" *ngFor="let user of filteredUsers" class="user-card" [user]="user" ></app-user-card>
+        <app-user-card [simple]="true" *ngFor="let user of this.filteredUsers" class="user-card" [user]="user" ></app-user-card>
       </div>
 
       <!-- List view -->
